Added ApowB

+++ src/test/java/day18_Function2/WAP02_ApowB.java
@@ -4,7 +4,7 @@ public class WAP02_ApowB {
 
 	public static void main(String[] args) {
 		
-		double ans = Math.pow(2, 3);
+		int ans = (int)Math.pow(2, 3);
 		System.out.println(ans);
 
 	}
